LayoutTests/platform/glib/imported/w3c/web-platform-tests/reporting/document-reporting-not-batch-different-document.https-expected.txt



FAIL Reports are not batched for same url in different document. assert_equals: Number of reports expected 2 but got 1
